new commit

+++ mock.py
@@ -6,6 +6,7 @@ import csv
 from sensormodel import getModel
 from androidEmulator import send
 import glob
+import os
 
 parser = argparse.ArgumentParser()
 parser.add_argument("magnitude", choices=("Lower", "Normal", "Higher"))
@@ -17,18 +18,22 @@ model = getModel(args.magnitude)
 if(args.frequency == 10000): period = None
 else: period = 1.0 / args.frequency # in seconds
 
-files = glob.glob(f"../pythonLogs/{args.magnitude}_{args.frequency}_{args.delay}_send_*.csv")
-iteration = len(files) + 1
-logFile = "../pythonLogs/{args.magnitude}_{args.frequency}_{args.delay}_send_{iteration}.csv"
-
+files = glob.glob(f"../pythonLogs/{args.magnitude}_{int(args.frequency)}_{args.delay}_send_*.csv")
+iteration = len(files)
+logFile = f"../pythonLogs/{args.magnitude}_{int(args.frequency)}_{args.delay}_send_{iteration}.csv"
 print("writing to ", logFile)
+
+if os.exists(logFile):
+	print("error in iteration numbers, exiting")
+	exit(1)
+
 
 t0 = time.time()
 now = t0
 end = t0 + 10.0
 with open(logFile, "w", newline="") as f:
 	writer = csv.writer(f)
-	writer.writerow(["Timestamp", "ax", "ay", "az"])
+	writer.writerow(["timestamp", "ax", "ay", "az"])
 
 	while end > now:
 		[ax, ay, az] = model.value(now-t0)
